Fix: 힌트 drawer 외부 클릭 핸들러 컴포넌트 내부로 수정
