+++ src/pages/CookieRecordPage/index.tsx
@@ -13,7 +13,6 @@ import {
 } from '@chakra-ui/react'
 
 import { Loading } from '@/components/Loading'
-import { useClickOutSideElement } from '@/hooks/useClickOutsideElement'
 
 import { CalendarSection } from './CalendarSection'
 import { CookieRecordErrorFallback } from './CookieRecordErrorFallback'
@@ -38,15 +37,29 @@ export default function CookieRecordPage() {
   const hintDrawer = useDisclosure()
   const hintModal = useDisclosure()
 
-  useClickOutSideElement(
-    document.getElementById('hint-drawer'),
-    hintDrawer.onClose,
-    hintModal.isOpen
-  )
-
   useEffect(() => {
     setPortalNode(document.getElementById('page-layout'))
   }, [])
+
+  useEffect(() => {
+    const element = document.getElementById('hint-drawer')
+
+    const listener = (event: MouseEvent) => {
+      if (
+        !hintModal.isOpen &&
+        element &&
+        !element.contains(event.target as Node)
+      ) {
+        hintDrawer.onClose()
+      }
+    }
+
+    document.addEventListener('mousedown', listener)
+
+    return () => {
+      document.removeEventListener('mousedown', listener)
+    }
+  }, [hintDrawer, hintModal])
 
   return (
     <Flex flexDirection="column">
